Prebid 9: Removing innerText & adding eslint rule (#11531)
* 11233 Removing innerText & adding eslint rule

* tests fix

---------
